// src/main/webapp/src/api/models/Person.ts
export interface Person {
    /**
     * 
     * @type {string}
     * @memberof Person
     */
    name?: string;
    /**
     * 
     * @type {PersonType}
     * @memberof Person
     */
    personType?: PersonType;
    /**
     * 
     * @type {Array<Skill>}
     * @memberof Person
     */
    skills?: Array<Skill>;
    /**
     * 
     * @type {Location}
     * @memberof Person
     */
    location?: Location;
    /**
     * 
     * @type {Array<TimeSlot>}
     * @memberof Person
     */
    availability?: Array<TimeSlot>;
    /**
     * 
     * @type {Array<Skill>}
     * @memberof Person
     */
    requiredSkills?: Array<Skill>;
    /**
     * 
     * @type {boolean}
     * @memberof Person
     */
    needsEvaluation?: boolean;
    /**
     * 
     * @type {Array<Person>}
     * @memberof Person
     */
    vetoes?: Array<Person>;
    /**
     * 
     * @type {Array<Array<string>>}
     * @memberof Person
     */
    hasAlreadyInspected?: Array<Array<string>>;
    /**
     * 
     * @type {number}
     * @memberof Person
     */
    maxNumberOfInspections?: number;
    /**
     * 
     * @type {Settings}
     * @memberof Person
     */
    settings?: Settings;
}